refactor: get text from soup
