+++ gpt_researcher/scraper/utils.py
@@ -2,6 +2,8 @@ from bs4 import BeautifulSoup
 from urllib.parse import urljoin, urlparse, parse_qs
 import logging
 import hashlib
+import re
+import bs4
 
 def get_relevant_images(soup: BeautifulSoup, url: str) -> list:
     """Extract relevant images from the page"""
@@ -84,3 +86,44 @@ def get_image_hash(image_url: str) -> str:
     except Exception as e:
         logging.error(f"Error calculating image hash for {image_url}: {e}")
         return None
+
+
+def clean_soup(soup: BeautifulSoup) -> BeautifulSoup:
+    """Clean the soup by removing unwanted tags"""
+    for tag in soup.find_all(
+        [
+            "script",
+            "style",
+            "footer",
+            "header",
+            "nav",
+            "menu",
+            "sidebar",
+            "svg",
+        ]
+    ):
+        tag.decompose()
+
+    disallowed_class_set = {"nav", "menu", "sidebar", "footer"}
+
+    # clean tags with certain classes
+    def does_tag_have_disallowed_class(elem) -> bool:
+        if not isinstance(elem, bs4.Tag):
+            return False
+
+        return any(
+            cls_name in disallowed_class_set for cls_name in elem.get("class", [])
+        )
+
+    for tag in soup.find_all(does_tag_have_disallowed_class):
+        tag.decompose()
+
+    return soup
+
+
+def get_text_from_soup(soup: BeautifulSoup) -> str:
+    """Get the relevant text from the soup with improved filtering"""
+    text = soup.get_text(strip=True, separator="\n")
+    # Remove excess whitespace
+    text = re.sub(r"\s{2,}", " ", text)
+    return text
